+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Removed
 - Deleted setting of the path `marfatech_rabbit_queue.connections`.
+- Provided message context while rewind that from queue by `QueueHeaderOptionEnum::X_CONTEXT`.
+- Argument `deliveryTagContextList` in rewind exceptions for transfer message context.
 
 ## [3.1.0] - 2022-05-16
 ### Added